fix: store templates bot & added Navigate Create Bot Template
