+++ src/services/types/bot.ts
@@ -51,6 +51,6 @@ export interface IGetBotsResponse extends Array<TBot> {}
 
 export interface IGetTemplatesBotsResponse extends Array<TTemplateBotRes> {}
 
-export interface IAddTemplatesBotsResponse extends Array<TTemplateBotRes> {}
+export interface IAddTemplatesBotsResponse extends TTemplateBotRes {}
 
 export interface IDeleteTemplatesBotsResponse extends TTemplateBotRes {}
